🎨 recording button 분리
